Move Custom Color serialization to its own class.
