+++ com.unity.shadergraph/Editor/Drawing/Colors/CustomColorData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 77e837a1c5a3a4f3f873a99e6287a220
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
